django/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
ippanel==2.0.7
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
pytest==7.2.0
pytest-django==4.5.2
Pillow==9.3.0